实现收益界面选择产品，图表显示对应产品的7天收益记录 Formatter类中添加了根据指定Date日期计算前n天Date的方法（只精确到天） 联系我们界面添加了让用户选择文字复制邮箱和号码的功能

+++ app/src/main/java/com/shenkong/bzzmaster/ui/activity/submit/SubmitOrderActivity.java
@@ -50,8 +50,8 @@ public class SubmitOrderActivity extends BaseMvpActivity<SubmitOrderPresenter> i
     private MaterialButton btnRechargeImmediately;
     private MaterialButton btnSubmitOrder;
 
-    private int count = 1;
-    private int minCount = 0;
+    private long count = 1;
+    private long minCount = 0;
     private ProductPlanBean productPlanBean;
     private androidx.core.widget.ContentLoadingProgressBar progress;
